*add doctors schedule table

See merge request practice-summer-2019/java/hospital/frontend!11

+++ src/app/lists/specialty-list/specialty-list.component.ts
@@ -12,10 +12,10 @@ export class SpecialtyListComponent implements OnInit {
 
   displayedColumns: string[] = ['specialty', 'delete'];
   dataSource;
-
-  constructor(private route: ActivatedRoute, private router: Router, private specialtyService: SpecialtyService) {}
-
   @ViewChild(MatSort, {static: true}) sort: MatSort;
+
+  constructor(private route: ActivatedRoute, private router: Router, private specialtyService: SpecialtyService) {
+  }
 
   ngOnInit() {
     this.reloadData();
@@ -29,13 +29,14 @@ export class SpecialtyListComponent implements OnInit {
   }
 
   deleteSpecialty(id: number) {
-    this.specialtyService.deleteSpecialty(id)
+    if (confirm('Уверены что хотите удалить?')) {
+      this.specialtyService.deleteSpecialty(id)
       .subscribe(
-        data => {
-          console.log(data);
-          this.reloadData();
-        },
-        error => console.log(error));
+          data => {
+            console.log(data);
+            this.reloadData();
+          },
+          error => console.log(error));
+    }
   }
-
 }
